modules/location-producer-service/requirements.txt
protobuf==4.21.6
grpcio==1.49.0
grpcio-tools==1.49.0
kafka-python==2.0.2